+++ __tests__2022/components/AwaitingApproval.test.tsx
@@ -1,21 +1,20 @@
 import React from 'react';
-import { renderHook } from '@testing-library/react-hooks';
-import { render, fireEvent, screen, cleanup } from '@testing-library/react';
-
+import { render, fireEvent, screen } from '@testing-library/react';
 import AwaitingApproval from '../../app/components/AwaitingApproval';
-
 import '@testing-library/jest-dom';
 
+jest.mock('react-router', () => ({
+  // ...jest.requireActual('react-router-dom') as typeof ReactRouterDom,
+  useHistory: () => ({ push: jest.fn() }), // ({ push: jest.fn() })
+}));
+
 describe('Awaiting Approval Page', () => {
-  const { result } = renderHook(() => AwaitingApproval());
-  
-  it('blah', () => {
-    expect(result.current).toBe('function');
+  beforeEach(() => {
+    render(<AwaitingApproval />);
   });
 
   it('Should have awaiting approval message', () => {
-    // screen.logTestingPlaygroundURL();
-    const element = result.current;
+    const element = screen.getByTestId('awaitingApprovalMessage');
     expect(element).toMatchInlineSnapshot(`
     <p
       class="welcomeMessage"
@@ -25,40 +24,15 @@ describe('Awaiting Approval Page', () => {
     </p>
     `);
   });
+
+  it('Should have return button', () => {
+    const returnBtn = screen.getByRole('button');
+    expect(returnBtn).toHaveTextContent('Return');
+  });
+
+  it('Should perform re-route on button click', () => {
+    const returnBtn = screen.getByRole('button');
+    expect(fireEvent.click(returnBtn)).toBeTruthy();
+    expect(typeof returnBtn.onclick).toBe('function');
+  });
 });
-// describe('Awaiting Approval Page', () => {
-//   beforeEach(() => {
-//     render(<AwaitingApproval />);
-//   });
-
-//   xit('Should have awaiting approval message', () => {
-//     const element = screen.getByTestId('awaitingApprovalMessage');
-//     expect(element).toMatchInlineSnapshot(`
-//     <p
-//       class="welcomeMessage"
-//       data-testid="awaitingApprovalMessage"
-//     >
-//       Your account is awaiting approval. Please contact your administrator if you have any questions.
-//     </p>
-//     `);
-//   });
-
-//   xit('Should have return button', () => {
-//     const returnBtn = screen.getByRole('button');
-//     expect(returnBtn).toHaveTextContent('Return');
-//   });
-
-//   it('Should perform re-route on button click', () => {
-//     const returnBtn = screen.getByRole('button');
-//     fireEvent.click(returnBtn);
-//     screen.logTestingPlaygroundURL();
-//     // expect(mockCall).toHaveBeenCalledTimes(1);
-//   });
-// });
-
-
-/* render(<AwaitingApproval onClick={mockCall() />})
-
-
-
-*/
