--- __tests__2022/test_users.json
@@ -1,1 +1,0 @@
-{"me@gmail.com":{"email":"me@gmail.com","username":"me","password":"1234","admin":true,"awaitingApproval":false},"you@gmail.com":{"email":"you@gmail.com","username":"you","password":"1234","admin":false,"awaitingApproval":false}}
